refactor: 컴포넌트 디렉토리구조 리팩토링 #67
