+++ client/src/components/molecules/Detail.tsx
