+++ client/src/components/organisms/SideBar.tsx
@@ -1,16 +1,21 @@
-import React, { ReactNode } from 'react';
+import React from 'react';
 import styled from '@emotion/styled';
+import Header from './Header';
+import DetailBackIcon from '../atoms/DetailBackIcon';
 import DateInfo from '../molecules/DateInfo';
-import Header from '../molecules/Header';
+import Detail from '../molecules/Detail';
+import PlaceList from '../molecules/PlaceList';
 import SearchContainer from '../molecules/SearchContainer';
+import { useQueryString } from '@/hooks/useQueryString';
 import BREAK_POINT from '@/styles/breakpoint';
 import { subtitle1 } from '@/styles/font';
 
 interface SideBarProps {
   title?: string;
-  children: ReactNode;
+  isLoading: boolean;
 }
-export default function SideBar({ title = '주변 장소', children }: SideBarProps) {
+export default function SideBar({ title = '주변 장소', isLoading }: SideBarProps) {
+  const { id } = useQueryString();
   return (
     <Wrapper>
       <Header />
@@ -20,7 +25,15 @@ export default function SideBar({ title = '주변 장소', children }: SideBarPr
       </DateContainer>
       <PlaceContainer>
         <Title>{title}</Title>
-        <Contents>{children}</Contents>
+        <Contents>
+          <PlaceList {...{ isLoading }} />
+          {id && (
+            <DetailWrapper>
+              <DetailBackIcon />
+              <Detail />
+            </DetailWrapper>
+          )}
+        </Contents>
       </PlaceContainer>
     </Wrapper>
   );
@@ -66,4 +79,20 @@ const Contents = styled.div`
   flex-direction: column;
   align-items: center;
   overflow-y: auto;
+`;
+
+const DetailWrapper = styled.div`
+  padding: 5px;
+  width: 390px;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  position: absolute;
+  left: 390px;
+  top: 0px;
+  z-index: 1;
+  background-color: ${({ theme }) => theme.color.white};
+  overflow-y: auto;
+  border-right: 1px solid ${({ theme }) => theme.color.gray200};
 `;
